+++ client/src/main.tsx
@@ -7,7 +7,7 @@ import { RouterProvider, createBrowserRouter } from "react-router-dom";
 
 // Import the main app component
 import App from "./App";
-import OfferNew from "./pages/offerNew/OfferNew";
+import OfferNew from "./pages/offerCreat/OfferCreat";
 
 // Import additional components for new routes
 // Try creating these components in the "pages" folder
@@ -23,11 +23,14 @@ const router = createBrowserRouter([
   {
     path: "/", // The root path
     element: <App />, // Renders the App component for the home page
+    children: [
+      {
+        path: "/offers/creat",
+        element: <OfferNew />,
+      },
+    ],
   },
-  {
-    path: "/offers/new",
-    element: <OfferNew />,
-  },
+
   // Try adding a new route! For example, "/about" with an About component
 ]);
